update deps, including clap to 3.2
This was mostly a test of the new value_parser APIs in clap 3.2, which
will be required in 4.0. The first couple 3.2.y releases were
frustrating, but things things quickly improved and I'm starting to see
the benefit now.

In this commit I could remove the parse_duration() function and inline
it into the clap derive attribute. Here's how this works:

  * #[clap(value_parser = X)] is equivalent to arg.value_parser(X)
  * Arg::value_parser takes an Into<clap::builder::ValueParser> argument
  * ValueParser implements From<P> where P: clap::builder::TypedValueParser
    (plus some extra send/sync/clone/'static trait bounds)
  * A function or closure which takes a &str and returns a Result
    implements TypedValueParser (modulo some Clone+Send+Sync+'static
    trait bounds)

Therefore I can just inline the value_parser as a closure and it all
works nicely, yay. This ends up being better than the old code because
I don't need a standalone parse_duration function halfway across the
file. Only "uglyness" is that I have to annotate |s: &str| in the
closure for some reason.

Also default_value_t would be nice to use for the timeout duration
rather than default_value, but it fails because std::time::Duration
doesn't implement Display.